\name{RSeqMeth-package}
\alias{RSeqMeth-package}
\alias{RSeqMeth}
\docType{package}
\title{
Package for analysis of Sequenom EpiTYPER Data
}
\description{
Package for analysis of Sequenom EpiTyper Data
}
\details{
\tabular{ll}{
Package: \tab RSeqMeth\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2007-08-13\cr
License: \tab GPL (Version 2 or later)\cr
}

}
\author{
Aaron Statham <a.statham@garvan.org.au>
}
\references{Coolen MW, Statham AL, Gardiner-Garden M, Clark SJ. (2007) Nucleic Acids Research 35(18):e119 Genomic profiling of CpG methylation and allelic specificity using quantitative high-throughput mass spectrometry: critical evaluation and improvements.}

\keyword{ package }

\examples{
#ampliconReport()

#analyzeSequenom()
}
